requirements.txt
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
tensorflow_hub==0.16.1
tf_keras==2.16.0
streamlit